Stepping Stone has basic functionality.
Notes:
====================================
I have checked in a Makefile and the stepping stone code for now.  Currently, it functions in a bare state.  There is no multi threading and I have added a few spots where people can integrate their work.  I added a test mode right now to test the stepping stone.  Here is how to run it right now in the test state:

1. Start four instances of the stepping stone, on ports 30030, 30031, 30032, 30033.  These four servers will just be waiting for an incoming message.
2. Then start another instance of the stepping stone on port 17 (run: ./ss -p 17).  This will execute a call to the server on port 30033 with a chaindata struct with 4 links in them, pointing to the four servers started in step 1.
3. Right now I have it hard coded to just return a test string from the last stepping stone.

I have included some points in the code where I think you can integrate your section of code.  Search for the following expressions with the parenthesis:

	(CLEANUP) - I have marked the spot(s) where I think the clean up code should be relevant.

	(URL HANDLER) - Right now I have just hard coded a return string, so there is a gap for the URL Handler to do its business.

	(SSHANDLER) - I marked a spot for where the random SS needs to be selected.  Currently, I just select the last SS from the end of the array.

ONE ISSUE - we do not have anyone assigned to code the awget module that actually connects to the initial (randomly selected) SS.  Can someone volunteer to do this?  If we are all too slammed, we will see who has more time and maybe it's a joint effort.

The next step would be for us to integrate our code together and then hopefully fix bugs and conquer the multi threading.  Also, can everyone please populate the README file with their information on github.  And can we plan to work tonight (Friday) on the project and continue throughout the weekend until we are finished?  Can't wait to see what we do...let's get a solid program together!